Fixes #26 patient name being set to null by the anonymisation.
Code originally buried in the uploading wizard gui was performing the necessary setting of XNAT variables that were later used in the uploading as part of processing the anonymisation script.
This code was no longer being called with the new uploader, so the variables project, subject and session were not being set and defaulted as null.

This commit fixes the project, subject and session variables. It does NOT fix the modalityLabel or other variables which were are not currently using.